requirements.txt
stream_chat==1.4.0
pydantic==1.6.1
fastapi==0.61.0
python-dotenv==0.14.0
uvicorn==0.11.8
